Refactor the reconciler table test logic so that the core is separable into knative/pkg.
