+++ pkg/reconciler/v1alpha1/route/table_test.go
@@ -1431,7 +1431,7 @@ func TestReconcile(t *testing.T) {
 	// TODO(mattmoor): Revision inactive (indirect reference)
 	// TODO(mattmoor): Multiple inactive Revisions
 
-	table.Test(t, func(listers *Listers, opt reconciler.Options) controller.Reconciler {
+	table.Test(t, MakeFactory(func(listers *Listers, opt reconciler.Options) controller.Reconciler {
 		return &Reconciler{
 			Base:                 reconciler.NewBase(opt, controllerAgentName),
 			routeLister:          listers.GetRouteLister(),
@@ -1448,7 +1448,7 @@ func TestReconcile(t *testing.T) {
 				},
 			},
 		}
-	})
+	}))
 }
 
 func mutateVirtualService(vs *istiov1alpha3.VirtualService) *istiov1alpha3.VirtualService {
